Logging and middleware refactoring
