+++ src/application.py
@@ -1,15 +1,11 @@
-import time
-
-import structlog
 from asgi_correlation_id import CorrelationIdMiddleware
-from asgi_correlation_id.context import correlation_id
-from fastapi import FastAPI, Request, Response
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from uvicorn.protocols.utils import get_path_with_query_string
 
 from src.api.router import category_router
 from src.bot.bot import start_bot
-from src.core.logging import setup_logging
+from src.core.logging.middleware import LoggingMiddleware
+from src.core.logging.setup import setup_logging
 from src.settings import settings
 
 
@@ -26,48 +22,7 @@ def create_app() -> FastAPI:
     )
 
     setup_logging()
-    access_logger = structlog.stdlib.get_logger("api.access")
-
-    @app.middleware("http")
-    async def logging_middleware(request: Request, call_next) -> Response:
-        """Настройка логирования для Uvicorn."""
-        structlog.contextvars.clear_contextvars()
-        request_id = correlation_id.get()
-        structlog.contextvars.bind_contextvars(request_id=request_id)
-
-        start_time = time.perf_counter_ns()
-        response = Response(status_code=500)
-        try:
-            response = await call_next(request)
-        except Exception:
-            structlog.stdlib.get_logger("api.error").exception(
-                "Непойманное исключение"
-            )
-            raise
-        finally:
-            process_time = time.perf_counter_ns() - start_time
-            status_code = response.status_code
-            url = get_path_with_query_string(request.scope)
-            client_host = request.client.host
-            client_port = request.client.port
-            http_method = request.method
-            http_version = request.scope["http_version"]
-            access_logger.info(
-                f'{client_host}:{client_port} - "{http_method} {url} '
-                f'HTTP/{http_version}" {status_code}',
-                http={
-                    "url": str(request.url),
-                    "status_code": status_code,
-                    "method": http_method,
-                    "request_id": request_id,
-                    "version": http_version,
-                },
-                network={"client": {"ip": client_host, "port": client_port}},
-                duration=process_time,
-            )
-            response.headers["X-Process-Time"] = str(process_time / 10**9)
-            return response
-
+    app.add_middleware(LoggingMiddleware)
     app.add_middleware(CorrelationIdMiddleware)
 
     app.include_router(router=category_router, prefix="/api")
@@ -75,7 +30,6 @@ def create_app() -> FastAPI:
     @app.on_event("startup")
     async def on_startup():
         """Действия при запуске сервера."""
-        pass
         bot_instance = await start_bot()
         # storing bot_instance to extra state of FastAPI app instance
         # refer to https://www.starlette.io/applications/#storing-state-on-the-app-instance
