+++ src/core/logging/middleware.py
@@ -1,0 +1,50 @@
+import time
+
+import structlog
+from asgi_correlation_id.context import correlation_id
+from fastapi import Request, Response
+from starlette.middleware.base import BaseHTTPMiddleware
+from uvicorn.protocols.utils import get_path_with_query_string
+
+access_logger = structlog.stdlib.get_logger("api.access")
+
+
+class LoggingMiddleware(BaseHTTPMiddleware):
+    async def dispatch(self, request: Request, call_next) -> Response:
+        """Настройка логирования для Uvicorn."""
+        structlog.contextvars.clear_contextvars()
+        request_id = correlation_id.get()
+        structlog.contextvars.bind_contextvars(request_id=request_id)
+
+        start_time = time.perf_counter_ns()
+        response = Response(status_code=500)
+        try:
+            response = await call_next(request)
+        except Exception as error:
+            structlog.stdlib.get_logger("api.error").exception(
+                "Непойманное исключение"
+            )
+            raise error
+        finally:
+            process_time = time.perf_counter_ns() - start_time
+            status_code = response.status_code
+            url = get_path_with_query_string(request.scope)
+            client_host = request.client.host
+            client_port = request.client.port
+            http_method = request.method
+            http_version = request.scope["http_version"]
+            access_logger.info(
+                f'{client_host}:{client_port} - "{http_method} {url} '
+                f'HTTP/{http_version}" {status_code}',
+                http={
+                    "url": str(request.url),
+                    "status_code": status_code,
+                    "method": http_method,
+                    "request_id": request_id,
+                    "version": http_version,
+                },
+                network={"client": {"ip": client_host, "port": client_port}},
+                duration=process_time,
+            )
+            response.headers["X-Process-Time"] = str(process_time / 10**9)
+            return response
