+++ src/run.py
@@ -1,5 +1,5 @@
 from src.bot import create_bot
-from src.core.logging import setup_logging
+from src.core.logging.setup import setup_logging
 
 if __name__ == "__main__":
     setup_logging()
