+++ src/settings.py
@@ -33,7 +33,7 @@ class Settings(BaseSettings):
     LOG_DIR: str | Path = BASE_DIR / "logs"
     LOG_FILE: str = "app.log"
     LOG_LEVEL: str = "INFO"
-    LOG_FILE_SIZE: int = 10 * 1024 * 1024
+    LOG_FILE_SIZE: int = 10 * 2**20
     LOG_FILES_TO_KEEP: int = 5
     LOG_TO_JSON: bool = False
